# Use `Uint8ClampedArray` when returning image data in `src/core/jbig2.js` and `src/core/jpg.js`